requirements.txt
streamlit
together
pytest
python-dotenv